Add files via upload
I have developed a game in which blocks are spawning, they will collide with the player  when player touches  with it. Also a life system is created in which stars and hearts are spawning, when player touches heart or star a count is shown on console . if health reduces and reaches -2 value ,the game will be over and its shown on console.